+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/notification/domain/base/service/NotificationPublisherService.java
@@ -58,14 +58,18 @@ public class NotificationPublisherService {
 
     public Notification startNotification(StartNotification startNotification) {
         BPN applicationBPN = traceabilityProperties.getBpn();
-        Notification notification = Notification.startNotification(startNotification.getTitle(), clock.instant(), applicationBPN, startNotification.getDescription(), startNotification.getType());
+        Notification notification = Notification.startNotification(startNotification.getTitle(), clock.instant(), applicationBPN, startNotification.getDescription(), startNotification.getType(), startNotification.getSeverity(), startNotification.getTargetDate());
         createMessages(startNotification, applicationBPN, notification, assetAsBuiltRepository);
         return notification;
     }
 
     private void createMessages(StartNotification startNotification, BPN applicationBPN, Notification notification, AssetAsBuiltRepository assetAsBuiltRepository) {
-        Map<String, List<AssetBase>> assetsAsBuiltBPNMap = assetAsBuiltRepository.getAssetsById(startNotification.getAffectedPartIds()).stream().collect(groupingBy(AssetBase::getManufacturerId));
-        assetsAsBuiltBPNMap
+        Map<String, List<AssetBase>> assetsAsBuiltBPNMap =
+                assetAsBuiltRepository
+                        .getAssetsById(startNotification.getAffectedPartIds())
+                        .stream()
+                        .filter(asset -> Objects.nonNull(asset.getManufacturerId()))
+                        .collect(groupingBy(AssetBase::getManufacturerId));        assetsAsBuiltBPNMap
                 .entrySet()
                 .stream()
                 .map(it -> {
@@ -75,8 +79,6 @@ public class NotificationPublisherService {
                             applicationBPN,
                             startNotification.getReceiverBpn(),
                             startNotification.getDescription(),
-                            startNotification.getTargetDate(),
-                            startNotification.getSeverity(),
                             startNotification.getType(),
                             it,
                             creator,
@@ -156,9 +158,9 @@ public class NotificationPublisherService {
         relevantNotifications.forEach(qNotification -> {
             switch (status) {
                 case ACKNOWLEDGED -> notification.acknowledge();
-                case ACCEPTED -> notification.accept(reason);
-                case DECLINED -> notification.decline(reason);
-                case CLOSED -> notification.close(reason);
+                case ACCEPTED -> notification.accept(reason, qNotification);
+                case DECLINED -> notification.decline(reason, qNotification);
+                case CLOSED -> notification.close(reason, qNotification);
                 default ->
                         throw new NotificationIllegalUpdate("Transition from status '%s' to status '%s' is not allowed for notification with id '%s'".formatted(notification.getNotificationStatus().name(), status, notification.getNotificationId()));
             }