+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/notification/application/alert/rest/NotificationControllerTest.java
@@ -39,7 +39,7 @@ import notification.request.StartNotificationRequest;
 import notification.request.UpdateNotificationStatusTransitionRequest;
 import notification.request.UpdateNotificationStatusRequest;
 import notification.response.NotificationIdResponse;
-import notification.response.NotificationReasonResponse;
+
 import notification.response.NotificationResponse;
 import notification.response.NotificationSeverityResponse;
 import notification.response.NotificationSideResponse;
@@ -107,20 +107,15 @@ class NotificationControllerTest {
                 .hasFieldOrPropertyWithValue("description", notification.getDescription())
                 .hasFieldOrPropertyWithValue("createdBy", notification.getNotifications().stream()
                         .findFirst()
-                        .map(NotificationMessage::getCreatedBy)
+                        .map(NotificationMessage::getSentBy)
                         .orElse(null))
                 .hasFieldOrPropertyWithValue("createdByName", notification.getNotifications().stream()
                         .findFirst()
-                        .map(NotificationMessage::getCreatedByName)
+                        .map(NotificationMessage::getSentByName)
                         .orElse(null))
                 .hasFieldOrPropertyWithValue("createdDate", notification.getCreatedAt().toString())
                 .hasFieldOrPropertyWithValue("assetIds", notification.getAffectedPartIds())
                 .hasFieldOrPropertyWithValue("channel", NotificationSideResponse.SENDER)
-                .hasFieldOrPropertyWithValue("reason", new NotificationReasonResponse(
-                        notification.getCloseReason(),
-                        notification.getAcceptReason(),
-                        notification.getDeclineReason()
-                ))
                 .hasFieldOrPropertyWithValue("sendTo", "recipientBPN")
                 .hasFieldOrPropertyWithValue("sendToName", "receiverManufacturerName")
                 .hasFieldOrPropertyWithValue("severity", NotificationSeverityResponse.MINOR);