+++ CHANGELOG.md
@@ -9,6 +9,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ## [UNRELEASED - DD.MM.YYYY]
 ### Changed
 - #965 Implement proxy functionality of the IRS policy store
+- #962 Changed notification model to new one in frontend/backend
 - XXX Updated sonar maven plugin to version 4.0.0.4121
 
 ### Added